Refactoring: Moved record to a top-level file
