+++ appserver/tests/application-tests/pom.xml
@@ -57,7 +57,7 @@
                 <type>zip</type>
                 <scope>test</scope>
             </dependency>
-            
+
             <dependency>
                 <groupId>org.jboss.arquillian</groupId>
                 <artifactId>arquillian-bom</artifactId>
@@ -71,14 +71,14 @@
                 <version>1.10.0.Final</version>
                 <scope>test</scope>
             </dependency>
-            
+
             <dependency>
                 <groupId>ee.omnifish.arquillian</groupId>
                 <artifactId>arquillian-glassfish-server-managed</artifactId>
                 <version>${omnifish.arquillian.version}</version>
                 <scope>test</scope>
             </dependency>
-            
+
             <dependency>
                 <groupId>org.jboss.shrinkwrap.resolver</groupId>
                 <artifactId>shrinkwrap-resolver-impl-maven</artifactId>
@@ -103,7 +103,7 @@
             </dependency>
         </dependencies>
     </dependencyManagement>
-    
+
     <dependencies>
         <dependency>
             <groupId>org.glassfish.main</groupId>
@@ -130,7 +130,7 @@
                 </plugin>
             </plugins>
         </pluginManagement>
-        
+
         <plugins>
             <!-- To disable bringing license files -->
             <plugin>
@@ -142,7 +142,7 @@
                      </execution>
                  </executions>
             </plugin>
-            
+
             <!-- Block our annotation processing to be inherited -->
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
@@ -157,14 +157,14 @@
                     </execution>
                 </executions>
             </plugin>
-            
+
             <!-- To disable usage of our default MANIFEST.MF -->
             <plugin>
                 <artifactId>maven-jar-plugin</artifactId>
                 <configuration combine.self="override">
                 </configuration>
             </plugin>
-            
+
             <!-- Deploy this pom, but no children except ... -->
             <plugin>
                 <artifactId>maven-deploy-plugin</artifactId>
